pyproject.toml
[tool.poetry]
name = "csinspect"
version = "1.0.0"
description = "a Twitter/X bot that replies to CS2 inspect links with a generated screenshot."
authors = ["Hexiro <mail@hexiro.me>"]
license = "GPL v3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tweepy = { extras = ["async"], version = "^4.14.0" }
httpx = "^0.25.0"
redis = "^5.0.1"
python-socketio = "^5.10.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
sentry-sdk = "^1.32.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
mypy = "^1.6.0"
isort = "^5.12.0"
black = "^23.10.0"
types-redis = "^4.6.0.7"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 121
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 121
combine_as_imports = true
combine_star = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
ignore_missing_imports = true

[tool.ruff]
select = ["ANN", "TCH", "RUF", "SIM", "N", "S", "B", "A", "C4", "EM", "INP", "PIE", "SIM", "ERA", "TRY", "TID", "F"]
ignore = ["A003"]  # attributes with the same name as builtins is okay  

[tool.ruff.flake8-annotations]
allow-star-arg-any = true
